pyproject.toml
[tool.poetry]
name = "crud-fastapi"
version = "0.1.0"
description = ""
authors = ["marcosviana91 <marcos.viana.91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
websockets = "^12.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
itsdangerous = "^2.1.2"
authlib = "^1.3.0"
sqlalchemy = "^2.0.29"
tinydb = "^4.8.0"
strawberry-graphql = "^0.221.1"
jinja2 = "^3.1.3"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
